Cleanup on file structure and linter compatibility
Clean up the file tree, add tests folder, separate routing with app/db connections. App.js in top folder now connects to db and should work with heroku deploy. server.js is in server folder for separation of concerns and closer directory access to routes, models, etc.
