+++ package.json
@@ -12,6 +12,7 @@
     "express": "^4.16.3",
     "helmet": "^3.12.1",
     "mongoose": "^5.2.3",
+    "mongoose-type-url": "^1.0.4",
     "path": "^0.12.7"
   }
 }
