--- server/users/employer/employees.js
@@ -1,6 +1,0 @@
-// Company Url
-// Industry
-// Description
-// Username
-// Password
-// Email
